refactor(daemon): extract startup failure log classification
Why:
- daemon startup probing mixed log reading with startup failure classification, which made the helper boundary unclear.

What Changed:
- keep Unity log I/O in DaemonStartupReadinessProbe and extract pure startup-log classification into DaemonStartupFailureLogClassifier.
- rename the probe helper to describe that it classifies the latest startup log text rather than resolving log I/O.
- update startup probe tests to construct the probe through the revised boundary.

Impact:
- startup failure diagnosis stays behaviorally the same while the probe has a clearer responsibility split.
